test(is-release-in-progress-v1): harden run tests per review
- Move shared PR fixtures into a dedicated fixtures.ts module so test
  files no longer import another *.test.ts (which can run as a side
  effect under node --test workers).
- Make the missing-token test async, await run(), and throw an Error
  rather than a plain object so the assertion deterministically reflects
  the failure path.
- Use forEach() for side-effect-only test-suite iteration.
