+++ .github/actions/is-release-in-progress-v1/src/isReleaseInProgress.test.ts
@@ -1,31 +1,12 @@
 import { describe, it } from 'node:test'
 import { strict as assert } from 'node:assert'
-import type { PullRequest } from './types'
 import isReleaseInProgress from './isReleaseInProgress.ts'
-
-export const BUG_PULL_REQUEST: PullRequest = {
-  id: 1,
-  state: 'open',
-  labels: [{ id: 12345, name: 'bug' }]
-}
-
-export const RELEASE_PULL_REQUEST: PullRequest = {
-  id: 2,
-  state: 'open',
-  labels: [{ id: 23456, name: 'release' }]
-}
-
-export const SIMILAR_PULL_REQUEST_1: PullRequest = {
-  id: 2,
-  state: 'open',
-  labels: [{ id: 34567, name: 'release-2023' }]
-}
-
-export const SIMILAR_PULL_REQUEST_2: PullRequest = {
-  id: 2,
-  state: 'open',
-  labels: [{ id: 45678, name: 'prerelease' }]
-}
+import {
+  BUG_PULL_REQUEST,
+  RELEASE_PULL_REQUEST,
+  SIMILAR_PULL_REQUEST_1,
+  SIMILAR_PULL_REQUEST_2
+} from './fixtures.ts'
 
 const testCases = [
   {
@@ -56,7 +37,7 @@ const testCases = [
 ]
 
 describe('isReleaseInProgress()', () => {
-  testCases.map(({ input, inputDescription, output }) => {
+  testCases.forEach(({ input, inputDescription, output }) => {
     it(`returns ${output} for ${inputDescription}`, () => {
       assert.strictEqual(isReleaseInProgress(input), output)
     })
